+++ product-service/src/handlers/getProductById.js
